Switch to int64 for provisional version counters to avoid overflow

+++ tree/transparency/algorithms/data_provider.go
@@ -254,6 +254,6 @@ func (dp *DataProvider) Output(n uint64, nP, m *uint64) (*structs.CombinedTreePr
 	return dp.handle.Output(positions, n, nP, m)
 }
 
-func (dp *DataProvider) StopCondition(x uint64, ver int) bool {
+func (dp *DataProvider) StopCondition(x uint64, ver int64) bool {
 	return dp.handle.StopCondition(x, ver)
 }
